+++ CHANGES.rst
@@ -4,7 +4,11 @@ Changelog
 1.0rc5 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Enable IDG-rosa theme and fix it tests
+  [jeanferri]
+
+- Add pink theme based on IDG-amarelo
+  [ramiroluz]
 
 
 1.0rc4 (2015-08-26)